refactor(services): migrate error_tracking to orchestra/error_tracking/
Move error tracking files to orchestra/error_tracking/:
- error_tracking_service.py → orchestra/error_tracking/service.py
- error_tracking_queries.py → orchestra/error_tracking/queries.py
- (cleanup.py was already migrated)

Created orchestra/error_tracking/__init__.py with ErrorTrackingService export.

Updated all imports:
- orchestra/error_tracking/service.py (queries imports)
- services/__init__.py (lazy import mapping)
